Move FarmClient.h and MinerAux.h from ethminer to eth
