+++ eth/MinerAux.h
