+++ eth/minerfarm.json
